refactor(matrix): 将 matrix 的扩展方法在 ext 命名空间抛出
